gdatetime: Clarify documentation about clamping on addition
If adding months or years to a date, the day of the month of the result
is supposed to be the same as in the input — but that doesn’t work if
that day doesn’t exist in the result month.

Clarify the documentation about what happens here (the day of the month
is clamped to the length of that month).

Signed-off-by: Philip Withnall <withnall@endlessm.com>
Reviewed-by: nobody